docs: more docs for next package
