+++ packages/core/src/react/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useFetchSearch';
 export * from './types';
 export * from './useFetchTerms';
 export * from './useFetchAuthorArchive';
+export * from './useFetchAppSettings';
