--- packages/next/src/blocks/ImageComponent.tsx
@@ -1,14 +1,0 @@
-import Image from 'next/image';
-
-import { ImageBlockProps } from '@10up/headless-core/react';
-
-export function ImageComponent({ src, alt, width, height, children }: ImageBlockProps) {
-	if (!width || !height) {
-		return children;
-	}
-	if (!src) {
-		return children;
-	}
-
-	return <Image src={src} alt={alt} width={width} height={height} layout="intrinsic" />;
-}
