+++ packages/next/src/blocks/index.ts
@@ -1,3 +1,2 @@
-export * from './ImageComponent';
 export * from './LinkBlock';
 export * from './TwitterBlock';
