+++ packages/next/src/components/Yoast.tsx
@@ -1,5 +1,13 @@
 import Head from 'next/head';
 
+/**
+ * The Yoast component renders the Yoast SEO meta tags.
+ * This component is automatically rendered by {@link HeadlessApp} so you don't have to manually render it.
+ *
+ * @param props Component props. Expects a single `seo` prop
+ *
+ * @category React Components
+ */
 export function Yoast({ seo }) {
 	return (
 		<Head>
