+++ packages/next/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './HeadlessApp';
 export * from './Yoast';
+export * from './ImageComponent';
