+++ packages/next/src/data/hooks/useAppSettings.ts
@@ -1,11 +1,5 @@
-import {
-	EndpointParams,
-	AppEntity,
-	AppSettingsStrategy,
-	getWPUrl,
-	FetchResponse,
-} from '@10up/headless-core';
-import { HookResponse, useFetch } from '@10up/headless-core/react';
+import { AppEntity, FetchResponse } from '@10up/headless-core';
+import { HookResponse, useFetchAppSettings } from '@10up/headless-core/react';
 import { SWRConfiguration } from 'swr';
 
 export interface useAppSettingsResponse extends HookResponse {
@@ -15,8 +9,28 @@ export interface useAppSettingsResponse extends HookResponse {
 /**
  * The useAppSettings hook
  *
- * @param params
- * @param options
+ * ## Usage
+ *
+ * ```tsx
+ * const { data, loading, error } = useAppSettings();
+ *
+ * // check loading and error states
+ * ```
+ *
+ * ### Server-Side-Rendering or Static-Site-Generation
+ *
+ * ```tsx
+ * export async function getServerSideProps(context) {
+ * 	const useAppSettingsData = await fetchHookData(useAppSettings.fetcher(), context);
+ * 	return addHookData([useAppSettingsData], {});
+ * }
+ * ```
+ *
+ * **Important**: You most likely want to fetch app settings on every route so
+ * that you can access global settings and menus in your pages & components
+ *
+ * @param params The parameters accepted by the hook
+ * @param options Options for the SWR configuration
  *
  * @category Data Fetching Hooks
  */
@@ -24,23 +38,7 @@ export function useAppSettings(
 	params = {},
 	options: SWRConfiguration<FetchResponse<AppEntity>> = {},
 ): useAppSettingsResponse {
-	const { data, error } = useFetch<AppEntity, EndpointParams>(
-		params,
-		useAppSettings.fetcher(),
-		options,
-	);
-
-	if (error) {
-		return { error, loading: false };
-	}
-
-	if (!data) {
-		return { loading: true };
-	}
-
-	const { result } = data;
-
-	return { data: result, loading: false };
+	return useFetchAppSettings(params, options);
 }
 
 /**
@@ -48,5 +46,5 @@ export function useAppSettings(
  */
 // eslint-disable-next-line no-redeclare
 export namespace useAppSettings {
-	export const fetcher = () => new AppSettingsStrategy(getWPUrl());
+	export const { fetcher } = useFetchAppSettings;
 }
