+++ packages/next/src/data/utils.ts
@@ -11,8 +11,18 @@ import {
 import { getHeadlessConfig } from '@10up/headless-core/utils';
 import { GetServerSidePropsContext, GetServerSidePropsResult, GetStaticPropsContext } from 'next';
 
+/**
+ * The supported options for {@link fetchHookData}
+ */
 export interface FetchHookDataOptions {
+	/**
+	 * This should match params passed to the hook on the client side.
+	 */
 	params?: any;
+
+	/**
+	 * Optional. If set, the data will be filtered given {@link FilterDataOptions}
+	 */
 	filterData?: FilterDataOptions;
 }
 
@@ -21,7 +31,7 @@ export interface FetchHookDataOptions {
  *
  * @param args - Array of catch-all arguments
  *
- * @returns path
+ * @category Next.js Data Fetching Utilities
  */
 export function convertToPath(args: string[] | undefined) {
 	if (!args) {
@@ -31,6 +41,35 @@ export function convertToPath(args: string[] | undefined) {
 	return `/${args.join('/')}`;
 }
 
+/**
+ * A function that implementeds data fetching on the server. This should be used in `getServerSideProps`
+ * or `getStaticProps`.
+ *
+ * Data fetching will be perfomed by the specified strategy and URL params will be automatically extracted
+ * from `context
+ *
+ * ## Usage
+ *
+ * ```ts
+ * export async function getServerSideProps(context) {
+ *	try {
+ * 		const usePostsHook = await fetchHookData(usePosts.fetcher(),context);
+ *
+ *		return addHookData([usePostsHook], {});
+ *	} catch (e) {
+ *		return handleError(e, context);
+ *	}
+ * }
+ * ```
+ *
+ * @param fetchStrategy The fetch strategy to use. Typically this is exposed by the hook e.g: `usePosts.fetcher()`
+ * @param ctx The Next.js context, either the one from `getServerSideProps` or `getStaticProps`
+ * @param options See {@link FetchHookDataOptions}
+ *
+ * @returns An object with a key of `data` and a value of the fetched data.
+ *
+ * @category Next.js Data Fetching Utilities
+ */
 export async function fetchHookData(
 	fetchStrategy: AbstractFetchStrategy<Entity, EndpointParams>,
 	ctx: GetServerSidePropsContext | GetStaticPropsContext,
@@ -76,11 +115,34 @@ export async function fetchHookData(
 	return { key: endpointUrlForKey, data: fetchStrategy.filterData(data, filterDataOptions) };
 }
 
-type HookState = {
+export type HookState = {
 	key: string;
 	data: FetchResponse<{ yoast_head_json: {}; yoast_head: {} }>;
 };
 
+/**
+ * The `addHookData` function is responsible for collecting all of the results from the `fetchHookData` function calls
+ * and prepares the shape of the data to match what the frameworks expects (such as setting initial values for SWR and collecting SEO data).
+ *
+ * ## Usage
+ *
+ * ```ts
+ * export async function getServerSideProps(context) {
+ *	try {
+ * 		const usePostsHook = await fetchHookData(usePosts.fetcher(),context);
+ *		const useAppSettingsHook = await fetchHookData(useAppSettings.fetcher(),context);
+ *		return addHookData([usePostsHook, useAppSettingsHook], {});
+ *	} catch (e) {
+ *		return handleError(e, context);
+ *	}
+ * }
+ * ```
+ *
+ * @param hookStates An array of resolved promises from {@link fetchHookData}
+ * @param nextProps Any additional props to pass to Next.js page routes.
+ *
+ * @category Next.js Data Fetching Utilities
+ */
 export function addHookData(hookStates: HookState[], nextProps) {
 	const { props = {}, ...rest } = nextProps;
 	const fallback = {};
@@ -116,6 +178,33 @@ function isStringArray(el): el is string[] {
 	return Array.isArray(el);
 }
 
+/**
+ * The `handleError` function is responsible for handling errors that occur during
+ * data fetching in `getServerSideProps` or `getStaticProps`.
+ *
+ * It also handles redirects if `redirectStrategy` is set to `404` in `headless.config.js`
+ *
+ * If `error` is of type {@link NotFoundError} it will redirect to the 404 page. Otherwise it will
+ * return a server error (500) page
+ * ## Usage
+ *
+ * ```ts
+ * export async function getServerSideProps(context) {
+ *	try {
+ * 		const usePostsHook = await fetchHookData(usePosts.fetcher(),context);
+ *		return addHookData([usePostsHook], {});
+ *	} catch (e) {
+ *		return handleError(e, context);
+ *	}
+ * }
+ * ```
+ *
+ * @param error The error object
+ * @param ctx The Next.js context
+ * @param rootRoute The root route (deprecated/unnecessary). This needs to be revisited
+ *
+ * @category Next.js Data Fetching Utilities
+ */
 export async function handleError(
 	error: Error,
 	ctx: GetServerSidePropsContext,
